+++ docs/src/components/ComponentDoc/ComponentDoc.tsx
@@ -1,6 +1,16 @@
 import * as React from 'react'
 import { RouteComponentProps, withRouter } from 'react-router-dom'
-import { Flex, Header, Icon, Dropdown, Text, Grid, Menu, Box } from '@stardust-ui/react'
+import {
+  tabListBehavior,
+  Flex,
+  Header,
+  Icon,
+  Dropdown,
+  Text,
+  Grid,
+  Menu,
+  Box,
+} from '@stardust-ui/react'
 
 import { getFormattedHash } from 'docs/src/utils'
 import ComponentDocLinks from './ComponentDocLinks'
@@ -13,7 +23,6 @@ import ExampleContext from 'docs/src/context/ExampleContext'
 import ComponentPlayground from 'docs/src/components/ComponentPlayground/ComponentPlayground'
 import { ComponentInfo } from 'docs/src/types'
 import ComponentBestPractices from './ComponentBestPractices'
-import { tabListBehavior } from 'src/lib/accessibility'
 import * as _ from 'lodash'
 
 const exampleEndStyle: React.CSSProperties = {
@@ -30,20 +39,19 @@ type ComponentDocProps = {
 type ComponentDocState = {
   activePath: string
   currentTabIndex: number
-  defaultPropComponent: string
 }
 
 class ComponentDoc extends React.Component<ComponentDocProps, ComponentDocState> {
   state = {
     activePath: '',
-    defaultPropComponent: '',
+    propComponent: '',
     currentTabIndex: 0,
   }
 
   tabRegex = new RegExp(/[^\/]*$/)
 
-  getTabIndexOrRedirectToDefault(tab: string) {
-    const lowercaseTabs = _.map(this.props.tabs, tab => tab.toLowerCase())
+  getTabIndexOrRedirectToDefault(tab: string, tabs) {
+    const lowercaseTabs = _.map(tabs, tab => tab.toLowerCase())
     const index = lowercaseTabs.indexOf(tab)
     if (index === -1) {
       const { history, location } = this.props
@@ -60,24 +68,22 @@ class ComponentDoc extends React.Component<ComponentDocProps, ComponentDocState>
   }
 
   componentWillMount() {
-    const { history, location } = this.props
+    const { history, location, tabs } = this.props
     const tab = location.pathname.match(this.tabRegex)[0]
-    const tabIndex = this.getTabIndexOrRedirectToDefault(tab)
+    const tabIndex = this.getTabIndexOrRedirectToDefault(tab, tabs)
     this.setState({ currentTabIndex: tabIndex })
 
     if (location.hash) {
       const activePath = getFormattedHash(location.hash)
       history.replace({ ...history.location, hash: activePath })
       this.setState({ activePath })
-      if (this.props.tabs[tabIndex] === 'Props') {
-        this.setState({ defaultPropComponent: activePath })
-      }
     }
   }
 
-  componentWillReceiveProps({ info, location }) {
+  componentWillReceiveProps({ info, location, tabs }) {
     const tab = location.pathname.match(this.tabRegex)[0]
-    this.setState({ currentTabIndex: this.getTabIndexOrRedirectToDefault(tab) })
+    const tabIndex = this.getTabIndexOrRedirectToDefault(tab, tabs)
+    this.setState({ currentTabIndex: tabIndex })
 
     if (info.displayName !== this.props.info.displayName) {
       this.setState({ activePath: undefined })
@@ -109,11 +115,6 @@ class ComponentDoc extends React.Component<ComponentDocProps, ComponentDocState>
     this.setState({ currentTabIndex: newIndex })
   }
 
-  onPropComponentSelected = (e, props) => {
-    const { history } = this.props
-    history.push({ ...history.location, hash: props.value })
-  }
-
   render() {
     const getA11ySelectionMessage = {
       onAdd: item => `${item} has been selected.`,
@@ -142,7 +143,7 @@ class ComponentDoc extends React.Component<ComponentDocProps, ComponentDocState>
     }
 
     const { info, tabs } = this.props
-    const { activePath, currentTabIndex, defaultPropComponent } = this.state
+    const { activePath, currentTabIndex } = this.state
 
     return (
       <div style={{ padding: '20px' }}>
@@ -197,12 +198,7 @@ class ComponentDoc extends React.Component<ComponentDocProps, ComponentDocState>
         {this.getCurrentTabTitle() === 'Accessibility' && <ComponentDocAccessibility info={info} />}
 
         {this.getCurrentTabTitle() === 'Props' && (
-          <ComponentProps
-            displayName={info.displayName}
-            props={info.props}
-            defaultComponentProp={defaultPropComponent}
-            onPropComponentSelected={this.onPropComponentSelected}
-          />
+          <ComponentProps displayName={info.displayName} props={info.props} />
         )}
 
         {this.getCurrentTabTitle() === 'Definition' && (