+++ packages/react/src/components/Tree/TreeItem.tsx
@@ -1,3 +1,4 @@
+import { Accessibility, treeItemBehavior } from '@stardust-ui/accessibility'
 import * as customPropTypes from '@stardust-ui/react-proptypes'
 import * as _ from 'lodash'
 import * as PropTypes from 'prop-types'
@@ -5,8 +6,6 @@ import * as React from 'react'
 
 import { Ref } from '@stardust-ui/react-component-ref'
 import TreeTitle, { TreeTitleProps } from './TreeTitle'
-import { treeItemBehavior } from '../../lib/accessibility'
-import { Accessibility } from '../../lib/accessibility/types'
 import {
   UIComponent,
   childrenExist,
@@ -127,7 +126,7 @@ class TreeItem extends UIComponent<WithAsProp<TreeItemProps>, TreeItemState> {
 
   static defaultProps = {
     as: 'div',
-    accessibility: treeItemBehavior,
+    accessibility: treeItemBehavior as Accessibility,
   }
 
   state = {