+++ packages/react/src/lib/index.ts
@@ -11,11 +11,7 @@ export { default as stringLiteralsArray } from './stringLiteralsArray'
 export { default as getOrGenerateIdFromShorthand } from './getOrGenerateIdFromShorthand'
 
 export * from './factories'
-export { default as callable } from './callable'
 export { default as constants } from './constants'
-export { default as getClasses } from './getClasses'
-export { default as getElementType } from './getElementType'
-export { default as getUnhandledProps } from './getUnhandledProps'
 export { default as mergeThemes } from './mergeThemes'
 export { default as mergeProviderContexts } from './mergeProviderContexts'