+++ packages/react/src/lib/renderComponent.tsx
@@ -1,11 +1,21 @@
+import {
+  AccessibilityDefinition,
+  FocusZoneMode,
+  FocusZoneDefinition,
+  Accessibility,
+} from '@stardust-ui/accessibility'
+import {
+  callable,
+  FocusZone,
+  FocusZoneProps,
+  FOCUSZONE_WRAP_ATTRIBUTE,
+  getElementType,
+  getUnhandledProps,
+} from '@stardust-ui/react-bindings'
 import cx from 'classnames'
 import * as React from 'react'
 import * as _ from 'lodash'
 
-import callable from './callable'
-import getClasses from './getClasses'
-import getElementType from './getElementType'
-import getUnhandledProps from './getUnhandledProps'
 import logProviderMissingWarning from './providerMissingHandler'
 import {
   ComponentStyleFunctionParam,
@@ -18,17 +28,9 @@ import {
   ComponentSlotStylesInput,
 } from '../themes/types'
 import { Props, ProviderContextPrepared } from '../types'
-import {
-  AccessibilityDefinition,
-  FocusZoneMode,
-  FocusZoneDefinition,
-  Accessibility,
-} from './accessibility/types'
 import { ReactAccessibilityBehavior, AccessibilityActionHandlers } from './accessibility/reactTypes'
 import getKeyDownHandlers from './getKeyDownHandlers'
 import { emptyTheme, mergeComponentStyles, mergeComponentVariables } from './mergeThemes'
-import { FocusZoneProps, FocusZone } from './accessibility/FocusZone'
-import { FOCUSZONE_WRAP_ATTRIBUTE } from './accessibility/FocusZone/focusUtilities'
 import createAnimationStyles from './createAnimationStyles'
 import Debug, { isEnabled as isDebugEnabled } from './debug'
 
@@ -47,7 +49,6 @@ export type RenderComponentCallback<P> = (config: RenderResultConfig<P>) => any
 
 export interface RenderConfig<P> {
   className?: string
-  defaultProps?: { [key: string]: any }
   displayName: string
   handledProps: string[]
   props: PropsWithVarsAndStyles
@@ -63,6 +64,7 @@ const emptyBehavior: ReactAccessibilityBehavior = {
 }
 
 const getAccessibility = (
+  displayName: string,
   props: State & PropsWithVarsAndStyles & { accessibility?: Accessibility },
   actionHandlers: AccessibilityActionHandlers,
   isRtlEnabled: boolean,
@@ -75,6 +77,20 @@ const getAccessibility = (
 
   const definition: AccessibilityDefinition = accessibility(props)
   const keyHandlers = getKeyDownHandlers(actionHandlers, definition.keyActions, isRtlEnabled)
+
+  if (process.env.NODE_ENV !== 'production') {
+    // For the non-production builds we enable the runtime accessibility attributes validator.
+    // We're adding the data-aa-class attribute which is being consumed by the validator, the
+    // schema is located in @stardust-ui/ability-attributes package.
+    if (definition.attributes) {
+      const slotNames = Object.keys(definition.attributes)
+      slotNames.forEach(slotName => {
+        definition.attributes[slotName]['data-aa-class'] = `${displayName}${
+          slotName === 'root' ? '' : `__${slotName}`
+        }`
+      })
+    }
+  }
 
   return {
     ...emptyBehavior,
@@ -146,7 +162,6 @@ const renderComponent = <P extends {}>(
 ): React.ReactElement<P> => {
   const {
     className,
-    defaultProps,
     displayName,
     handledProps,
     props,
@@ -163,7 +178,7 @@ const renderComponent = <P extends {}>(
   const { disableAnimations = false, renderer = null, rtl = false, theme = emptyTheme } =
     context || {}
 
-  const ElementType = getElementType({ defaultProps }, props) as React.ReactType<P>
+  const ElementType = getElementType(props) as React.ReactType<P>
   const stateAndProps = { ...state, ...props }
 
   // Resolve variables for this component, allow props.variables to override
@@ -185,12 +200,13 @@ const renderComponent = <P extends {}>(
   )
 
   const accessibility: ReactAccessibilityBehavior = getAccessibility(
+    displayName,
     stateAndProps,
     actionHandlers,
     rtl,
   )
 
-  const unhandledProps = getUnhandledProps({ handledProps }, props)
+  const unhandledProps = getUnhandledProps(handledProps, props)
 
   const styleParam: ComponentStyleFunctionParam = {
     displayName,
@@ -201,11 +217,25 @@ const renderComponent = <P extends {}>(
     disableAnimations,
   }
 
-  const resolvedStyles: ComponentSlotStylesPrepared = resolveStyles(mergedStyles, styleParam)
+  // Fela plugins rely on `direction` param in `theme` prop instead of RTL
+  // Our API should be aligned with it
+  // Heads Up! Keep in sync with Design.tsx render logic
+  const direction = rtl ? 'rtl' : 'ltr'
+  const felaParam = {
+    theme: { direction },
+  }
 
-  const classes: ComponentSlotClasses = renderer
-    ? getClasses(renderer, mergedStyles, styleParam)
-    : {}
+  const resolvedStyles: ComponentSlotStylesPrepared = {}
+  const classes: ComponentSlotClasses = {}
+
+  Object.keys(mergedStyles).forEach(slotName => {
+    resolvedStyles[slotName] = callable(mergedStyles[slotName])(styleParam)
+
+    if (renderer) {
+      classes[slotName] = renderer.renderRule(callable(resolvedStyles[slotName]), felaParam)
+    }
+  })
+
   classes.root = cx(className, classes.root, props.className)
 
   const resolvedConfig: RenderResultConfig<P> = {