+++ packages/react/src/themes/teams/components/Icon/svg/icons/index.ts
@@ -9,6 +9,7 @@ import arrowLeft from './arrowLeft'
 import arrowRight from './arrowRight'
 import audioOff from './audioOff'
 import ban from './ban'
+import bell from './bell'
 import bellMute from './bellMute'
 import bold from './bold'
 import bookmark from './bookmark'
@@ -75,6 +76,7 @@ import filesPowerPointBrand from './filesPowerPointBrand'
 import filesWord from './filesWord'
 import filesWordBrand from './filesWordBrand'
 import filesVisio from './filesVisio'
+import filesVisioBrand from './filesVisioBrand'
 import fontColor from './fontColor'
 import fontSize from './fontSize'
 import format from './format'
@@ -100,6 +102,8 @@ import messageSeen from './messageSeen'
 import mic from './mic'
 import micOff from './micOff'
 import more from './more'
+import notes from './notes'
+import notificationsMuted from './notificationsMuted'
 import oneDrive from './oneDrive'
 import microsoftStream from './microsoftStream'
 import userPhone from './userPhone'
@@ -118,6 +122,7 @@ import poll from './poll'
 import presenceAvailable from './presenceAvailable'
 import presenceStroke from './presenceStroke'
 import qna from './qna'
+import questionCircle from './questionCircle'
 import quote from './quote'
 import readAloud from './read-aloud'
 import redbang from './redbang'
@@ -136,12 +141,15 @@ import star from './star'
 import stardustCircle from './stardustCircle'
 import sticker from './sticker'
 import strike from './strike'
+import sync from './sync'
 import table from './table'
 import tableAdd from './tableAdd'
 import tableDelete from './tableDelete'
 import teamCreate from './teamCreate'
 import teams from './teams'
 import toDoList from './toDoList'
+import thumbtack from './thumbtack'
+import thumbtackSlash from './thumbtackSlash'
 import translation from './translation'
 import trashCan from './trashCan'
 import triangleDown from './triangleDown'
@@ -168,7 +176,9 @@ export default {
   'arrow-right': arrowRight,
   'audio-off': audioOff,
   ban,
+  bell,
   'bell-slash': bellMute,
+  'bell-snooze': snooze,
   bold,
   bookmark,
   broadcast,
@@ -214,7 +224,6 @@ export default {
   'files-txt': filesTxt,
   'files-upload': filesUpload,
   'files-video': filesVideo,
-  'files-visio': filesVisio,
   'files-zip': filesZip,
   flag,
   download,
@@ -249,14 +258,14 @@ export default {
   menu,
   'message-seen': messageSeen,
   mic,
+  notes,
   'speaker-mute': speakerMute,
   hand,
   'mic-off': micOff,
   more,
-  'microsoft-stream': microsoftStream,
   'user-phone': userPhone,
+  'notifications-muted': notificationsMuted,
   'number-list': numberList,
-  onedrive: oneDrive,
   'open-outside': openOutside,
   outdent,
   paperclip,
@@ -271,6 +280,7 @@ export default {
   'presence-available': presenceAvailable,
   'presence-stroke': presenceStroke,
   qna,
+  'question-circle': questionCircle,
   quote,
   'read-aloud': readAloud,
   redbang,
@@ -278,19 +288,20 @@ export default {
   'remove-format': removeFormat,
   reply,
   retry,
-  'skype-logo': skypeLogo,
   search,
   send,
   settings,
   'share-alt': shareAlt,
-  'bell-snooze': snooze,
   star,
   sticker,
   strike,
+  sync,
   table,
   'table-add': tableAdd,
   'table-delete': tableDelete,
   teams,
+  thumbtack,
+  'thumbtack-slash': thumbtackSlash,
   'to-do-list': toDoList,
   translation,
   'trash-can': trashCan,
@@ -304,19 +315,23 @@ export default {
   'user-blur': userBlur,
   'user-friends': userFriends,
   'video-camera-emphasis': videoCameraEmphasis,
-  yammer,
   'window-maximize': windowMaximize,
   'window-minimize': windowMinimize,
   'window-restore': windowRestore,
 
   // Office icons
-  word: filesWord,
   excel: filesExcel,
-  powerpoint: filesPowerPoint,
-  onenote: filesOneNote,
-
-  'word-color': filesWordBrand,
   'excel-color': filesExcelBrand,
-  'powerpoint-color': filesPowerPointBrand,
+  onedrive: oneDrive,
+  onenote: filesOneNote,
   'onenote-color': filesOneNoteBrand,
+  powerpoint: filesPowerPoint,
+  'powerpoint-color': filesPowerPointBrand,
+  'skype-logo': skypeLogo,
+  'microsoft-stream': microsoftStream,
+  word: filesWord,
+  'word-color': filesWordBrand,
+  yammer,
+  visio: filesVisio,
+  'visio-color': filesVisioBrand,
 } as { [iconName: string]: TeamsSvgIconSpec }